we commence-- THE REBOOT OF CONFIGURATION.  see full message
Yeah, having done one, I see a way to automate the hell out of this.
the ConfigMenu class lets one interact with a single JSON config file.

It is not, as yet, supremely flexible, and I'm gonna do something about
that, but this is gonna be a big help moving forward.