added write functionality, more doc, minor changes (see notes)
writeCenter() has the potential to provide a negative x coordinate
which could potentially cause it to violate padding and/or border,
or even fail.  This has been resolved by moving the truncation
functionality that previously was in writeRight() into write(),
which is called by both writeRight() and writeCenter().

setCursor() did not account for offsets.  This has been fixed.

constructor previously botched the logic and would have caused
problems if width or height (but not both) was set to None.  This has
been fixed.

added default of True to setEcho.

added properties height and width to keep track of the dimensions
of the window, rather than the available area.

reopen() now checks current size vs height,width and
resizes the screen if they do not match. I... can't actually explain
the logic used before.